Try/catch for mnRoute.renderView
- Introduced mnRoute.hasRenderedView
- Moved test from events to render, isolated from `load` dependency
